AzureBatchPoolResource
