+++ service/src/test/java/bio/terra/workspace/app/controller/ControlledAzureResourceApiControllerBatchPoolTest.java
@@ -1,5 +1,7 @@
 package bio.terra.workspace.app.controller;
 
+import static bio.terra.workspace.common.fixtures.ControlledAzureResourceFixtures.createAzureBatchPoolResource;
+import static bio.terra.workspace.common.fixtures.ControlledAzureResourceFixtures.createBatchPoolWithRequiredParameters;
 import static bio.terra.workspace.common.fixtures.ControlledResourceFixtures.makeDefaultControlledResourceFieldsApi;
 import static bio.terra.workspace.common.utils.MockMvcUtils.CREATE_AZURE_BATCH_POOL_PATH_FORMAT;
 import static bio.terra.workspace.common.utils.MockMvcUtils.DEFAULT_USER_EMAIL;
@@ -12,7 +14,6 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import bio.terra.workspace.common.BaseAzureUnitTest;
-import bio.terra.workspace.common.fixtures.ControlledResourceBatchPoolFixtures;
 import bio.terra.workspace.generated.model.ApiAzureBatchPoolUserAssignedIdentity;
 import bio.terra.workspace.generated.model.ApiControlledResourceCommonFields;
 import bio.terra.workspace.generated.model.ApiCreateControlledAzureBatchPoolRequestBody;
@@ -61,8 +62,7 @@ public class ControlledAzureResourceApiControllerBatchPoolTest extends BaseAzure
   public void createBatchPool400WithInconsistentUAMI() throws Exception {
     final ApiControlledResourceCommonFields commonFields = makeDefaultControlledResourceFieldsApi();
 
-    var creationParameters =
-        ControlledResourceBatchPoolFixtures.createBatchPoolWithRequiredParameters();
+    var creationParameters = createBatchPoolWithRequiredParameters();
     // name and clientId are mutually exclusive
     creationParameters.userAssignedIdentities(
         List.of(
@@ -75,7 +75,7 @@ public class ControlledAzureResourceApiControllerBatchPoolTest extends BaseAzure
 
     UUID workspaceId = UUID.randomUUID();
     ControlledAzureBatchPoolResource resource =
-        ControlledResourceBatchPoolFixtures.createAzureBatchPoolResource(
+        createAzureBatchPoolResource(
             creationParameters,
             controller.toCommonFields(
                 workspaceId,
@@ -105,8 +105,7 @@ public class ControlledAzureResourceApiControllerBatchPoolTest extends BaseAzure
   public void createBatchPoolWithRequiredParametersSuccess() throws Exception {
     final ApiControlledResourceCommonFields commonFields = makeDefaultControlledResourceFieldsApi();
 
-    var creationParameters =
-        ControlledResourceBatchPoolFixtures.createBatchPoolWithRequiredParameters();
+    var creationParameters = createBatchPoolWithRequiredParameters();
 
     final ApiCreateControlledAzureBatchPoolRequestBody batchPoolRequest =
         new ApiCreateControlledAzureBatchPoolRequestBody()
@@ -116,7 +115,7 @@ public class ControlledAzureResourceApiControllerBatchPoolTest extends BaseAzure
     setupMockLandingZoneRegion(Region.GERMANY_CENTRAL);
     UUID workspaceId = UUID.randomUUID();
     ControlledAzureBatchPoolResource resource =
-        ControlledResourceBatchPoolFixtures.createAzureBatchPoolResource(
+        createAzureBatchPoolResource(
             creationParameters,
             controller.toCommonFields(
                 workspaceId,
